Add parts for Dual Alphanumeric 14-Segment Displays.
The breadboard view is extracted from the breadboard view of the
backpack version.  The PCB view is based on the 600-mil spacing 18-pin
mystery part, with a larger silkscreen rectangle to represent the size
of the display.  The schematic view is extracted from the datasheet on
Adafruit's site for the red version.  This admittedly doesn't quite fit
Fritzing's style, but I'm no SVG artist so it will have to do unless or
until somebody comes along with something better.